pyproject.toml
[tool.poetry]
name = "ddp_ingest"
version = "0.0.0"
description = ""
authors = ["Monster Dev <monsterdev@broadinstitute.org>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
dagster = "^0.11.12"
broad-dagster-utils = "^0.5.0"
dagit = "^0.11.12"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"
flake8 = "^3.8.4"
mypy = "^0.812"
pre-commit = "^2.11.0"
pytest = "^6.2.1"

[build-system]
requires = ["poetry-core=^1.1.5"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
aggressive = 1
exclude = ".pytest_cache,__pycache__"
max_line_length = 120
